requirements.txt
groq
python-dotenv
requests
tenacity